EH-747: Moved YhteisenTutkinnonOsanOsaAlue.ts to a new folder
